Moved AdminDataTableModule to components folder, removed common folder, and updated all imports
