--- projects/app/src/app/pages/admin/common/index.ts
@@ -1,1 +1,0 @@
-export * from './admin-data-table';
