+++ projects/app/src/app/pages/admin/components/index.ts
@@ -1,2 +1,3 @@
+export * from './admin-data-table';
 export * from './admin-edit';
 export * from './admin-manage';
