+++ projects/app/src/app/pages/admin/pages/admin-filter-models/admin-filter-models-manage-properties/admin-filter-models-manage-properties.component.ts
@@ -7,8 +7,7 @@ import { BehaviorSubject, from, map, Observable, skipWhile, Subject, takeUntil, 
 import { FilterModelProperty, FilterModelPropertyQuery, FilterModelPropertyService } from '@app/features';
 import { PaginationRequest } from '@core/models/pagination';
 import { ConfirmationAlertModalCompoonent } from '@shared/modals/confirmation-alert';
-import { AdminColumn } from '../../../common';
-import { IAdminManageView } from '../../../components';
+import { AdminColumn, IAdminManageView } from '../../../components';
 
 @Component({
   selector: 'app-admin-filter-models-manage-properties',
