+++ projects/app/src/app/pages/admin/pages/admin-filters/admin-filter-sections-parts-manage/admin-filter-sections-parts-manage.component.ts
@@ -7,8 +7,7 @@ import { BehaviorSubject, from, map, Observable, skipWhile, Subject, takeUntil, 
 import { FilterSectionPart, FilterSectionPartQuery, FilterSectionPartService } from '@app/features';
 import { PaginationRequest } from '@core/models/pagination';
 import { ConfirmationAlertModalCompoonent } from '@shared/modals/confirmation-alert';
-import { AdminColumn } from '../../../common';
-import { IAdminManageView } from '../../../components';
+import { AdminColumn, IAdminManageView } from '../../../components';
 
 @Component({
   selector: 'app-admin-filters-sections-parts-manage',
